Added model to database mapping in the model classes. Unless you have the database set up, you probably cannot run the app, so look in the sql_scripts folder and do what it says.
** IMPORTANT -> Whenever I make changes to MimirDatabaseScript.sql, that
means our database changed, so rerun the script everytime you see I make
a change. 

Some new models have been added to help map the database better. The
Book class was simplified so that Hibernate stopped complaining that I
wasn't mapping all the values. 

Oh Sign in is broken, for now, it might be working again soon, but it's
because I changed RegisteredUser and don't want to update the controller
with how to create a new one.