Updated database - fixed enum issue with model mappings.
Added ajax call to book modal when the format dropdown is changed, it
gets any data needed for the book to update the display. I didn't 
add any of the javascript to update these fields yet, and we need to
display those fields first. 

Fixed the return and borrow book dao methods to not increment or 
decrement a book with unlimited copies.